Broke Reveal classes into separate files, added appropriate extension of, y'know, `.js`
