+++ js/ui/motif.reveal-trigger.es6.js
@@ -19,6 +19,9 @@
  * @author Jonathan Pacheco <jonathan@lifeblue.com>
  */
 
+import getAnimationFrame from "../utils/motif.animationFrame.es6";
+import {customEvent} from "../utils/motif.events.es6";
+
 const DEFAULTS = {
     trigger: "click",
     type: "default",
@@ -27,9 +30,6 @@ const DEFAULTS = {
     // Classes
     activeClass: "is-current",
     visitedClass: "is-visited",
-
-    // Hover Intent
-    hoverIntent: null,
 
     // Callbacks
     onInit: null,
@@ -46,54 +46,9 @@ const FOR_ATTR = "data-reveal-for";
 const OPTIONS_ATTR = "data-reveal-options";
 
 let counter = 0;
-const animationFrame =
-    window.requestAnimationFrame ||
-    function(callback, elem) {
-        window.setTimeout(callback, 1000 / 60);
-    };
-const queueDoneEvent = customEvent("reveal/queue/done");
+const animationFrame = getAnimationFrame();
 
-export default class Reveal {
-    constructor(elems, userOptions) {
-        if (typeof elems === "string") {
-            elems = document.querySelectorAll(elems);
-        }
-        if (elems.length) {
-            elems = [].slice.call(elems);
-        } else {
-            elems = [elems];
-        }
-
-        this.revealTriggers = elems.map(elem => {
-            return new RevealTrigger(elem, userOptions);
-        });
-
-        this.executeQueue();
-    }
-
-    /**
-     * Execute the queue for this reveal group
-     */
-    executeQueue() {
-        const queue = window.Reveal.queue;
-
-        /**
-         * If there's anything in the queue, loop through it
-         * and trigger a show for each one
-         */
-        if (queue.length) {
-            queue.forEach(queueEvent => {
-                document.dispatchEvent(queueEvent);
-            });
-        }
-
-        // Clear queue
-        window.Reveal.queue = [];
-
-        document.dispatchEvent(queueDoneEvent);
-    }
-}
-export class RevealTrigger {
+export default class RevealTrigger {
     constructor(elem, userOptions) {
         // Init Vars
         this.elem = elem;
@@ -717,17 +672,6 @@ export class RevealTrigger {
 
         return this.reference;
     }
-}
-
-function customEvent(name, data) {
-    if (window.CustomEvent) {
-        return new CustomEvent(name, { detail: data });
-    }
-
-    const event = document.createEvent("CustomEvent");
-    event.initCustomEvent(name, true, true, data);
-
-    return event;
 }
 
 /*
