+++ js/utils/motif.animationFrame.es6.js
@@ -1,0 +1,6 @@
+export default function getAnimationFrame() {
+  return window.requestAnimationFrame ||
+    function (callback, elem) {
+      window.setTimeout(callback, 1000 / 60);
+    };
+}
